wishlist/requirements.txt
azure-functions
pyodbc==5.0.1